+++ src/Repository/DoctorRepository.java
@@ -8,7 +8,6 @@ import java.util.Date;
 
 public class DoctorRepository extends BaseRepository<Doctor> {
     private ArrayList<Doctor> doctorRepository;
-
     public DoctorRepository() {
         super();
         this.doctorRepository = new ArrayList<>();
@@ -17,19 +16,45 @@ public class DoctorRepository extends BaseRepository<Doctor> {
     public DoctorRepository(ArrayList<Doctor> doctors) {
         super();
         this.doctorRepository = doctors;
-
+    }
+    public Doctor findByID(int doctorID){
+        Doctor doctor = null;
+        for(Doctor doctor1: doctorRepository){
+            if(doctor1.getDoctorID() == doctorID){
+                doctor = doctor1;
+                break;
+            }
+        }
+        return doctor;
     }
 
-    public void enrollDoctor(Doctor doctor, Department department) {
+    public boolean enrollDoctor(int doctorID,int departmentID,DepartmentRepository departmentRepository) {
+        Doctor doctor = findByID(doctorID);
+        if(doctor == null){
+            return false;
+        }
+        Department department = departmentRepository.findByID(departmentID);
+        if(department == null)
+            return false;
+
         if (!doctor.getDepartments().contains(department))
             doctor.enrollInDepartment(department);
         if (!department.getDoctors().contains(doctor))
             department.enrollDoctor(doctor);
+        return true;
     }
 
-    public void removeFromDepartment(Doctor doctor, Department department) {
+    public boolean removeFromDepartment(int doctorID,int departmentID,DepartmentRepository departmentRepository) {
+        Doctor doctor = findByID(doctorID);
+        if(doctor == null){
+            return false;
+        }
+        Department department = departmentRepository.findByID(departmentID);
+        if(department == null)
+            return false;
         doctor.removeDepartment(department);
         department.removeDoctor(doctor);
+        return true;
     }
 
     @Override
@@ -38,8 +63,13 @@ public class DoctorRepository extends BaseRepository<Doctor> {
     }
 
     @Override
-    public boolean remove(Doctor doctor) {
-        return doctorRepository.remove(doctor);
+    public boolean remove(int id) {
+        for (Doctor doctor : doctorRepository) {
+            if (doctor.getDoctorID() == id) {
+                return doctorRepository.remove(doctor);
+            }
+        }
+        return false;
     }
 
     @Override
@@ -48,27 +78,72 @@ public class DoctorRepository extends BaseRepository<Doctor> {
     }
 
 
-    public void updateFirstName(Doctor doctor, String name) {
-        doctor.setFirstName(name);
+    public boolean updateFirstName(int ID, String name) {
+        for (Doctor doctor : doctorRepository) {
+            if (doctor.getDoctorID() == ID) {
+                doctor.setFirstName(name);
+                return true;
+            }
+        }
+        return false;
+
     }
 
-    public void updateLastName(Doctor doctor, String name) {
-        doctor.setLastName(name);
+    public boolean updateLastName(int ID, String name) {
+        for (Doctor doctor : doctorRepository) {
+            if (doctor.getDoctorID() == ID) {
+                doctor.setLastName(name);
+                return true;
+            }
+        }
+        return false;
+
     }
 
 
-    public void updateContact(Doctor doctor, String contact) {
-        doctor.setContact(contact);
+    public boolean updateContact(int ID, String contact) {
+        for (Doctor doctor : doctorRepository) {
+            if (doctor.getDoctorID() == ID) {
+                doctor.setContact(contact);
+                return true;
+            }
+        }
+        return false;
+
     }
 
-    public void updateAddress(Doctor doctor, String address) {
-        doctor.setAddress(address);
+    public boolean updateAddress(int ID, String address) {
+        for (Doctor doctor : doctorRepository) {
+            if (doctor.getDoctorID() == ID) {
+                doctor.setAddress(address);
+                return true;
+            }
+        }
+        return false;
+
     }
 
-    public void updateBirthDate(Doctor doctor, Date date) {
-        doctor.setBirthDate(date);
+    public boolean updateBirthDate(int ID, Date date) {
+        for (Doctor doctor : doctorRepository) {
+            if (doctor.getDoctorID() == ID) {
+                doctor.setBirthDate(date);
+                return true;
+            }
+        }
+        return false;
+
     }
-    public ArrayList<Department> getDepartments(Doctor doctor){
+    public ArrayList<Department> getDepartments(int doctorID){
+        Doctor doctor = null;
+        for(Doctor doctor1: doctorRepository){
+            if(doctor1.getDoctorID() == doctorID){
+                doctor = doctor1;
+                break;
+            }
+        }
+        if(doctor == null){
+            return new ArrayList<>();
+        }
         return doctor.getDepartments();
     }
 }
